+++ clientd3d/bitmap/logo.bmp
