--- module/intro/makefile
@@ -1,14 +1,0 @@
-# Makefile for intro DLL
-
-TOPDIR=..\..
-!include $(TOPDIR)\common.mak
-
-MODULE = intro
-
-!include $(MODULEDIR)\module.mak
-
-$(OUTDIR)\intro.dll: \
-        $(OUTDIR)\intro.obj \
-        $(OUTDIR)\intro.res
-
-!include $(TOPDIR)\rules.mak
