+++ module/makefile
@@ -3,7 +3,7 @@
 TOPDIR=..
 !include $(TOPDIR)\common.mak
 
-MODULES = admin char mailnews merintr intro dm chess
+MODULES = admin char mailnews merintr dm chess
 
 all: $(MODULES)
